moving menu to its own namespace
